Moved function quantity_allclose to astropy.units.quantity
