+++ CHANGES.rst
@@ -69,6 +69,11 @@ astropy.table
 astropy.tests
 ^^^^^^^^^^^^^
 
+- The function ``quantity_allclose`` was moved to the ``units`` package with the new,
+  shorter name ``allclose``. It is deprecated in ``tests``. [#7252]
+
+- The function ``_unquantify_allclose_arguments`` was moved to the ``units`` package. It is deprecated in ``tests``. [#7252]
+
 astropy.time
 ^^^^^^^^^^^^
 
@@ -78,6 +83,11 @@ astropy.time
 
 astropy.units
 ^^^^^^^^^^^^^
+
+-  Added a units-aware ``allclose`` function (this was previously available in the ``tests`` module as ``quantity_allclose``) [#7252]
+
+-  Added a units-aware ``_unquantify_allclose_arguments`` function (this was previously available in the ``tests`` module as
+	``_unquantify_allclose_arguments``) [#7252]
 
 astropy.utils
 ^^^^^^^^^^^^^
