+++ astropy/coordinates/tests/test_api_ape5.py
@@ -17,11 +17,11 @@ import numpy as np
 from numpy.random import randn
 from numpy import testing as npt
 
-from ...tests.helper import (raises, quantity_allclose as allclose,
-                             assert_quantity_allclose as assert_allclose)
+from ...tests.helper import (raises, assert_quantity_allclose as assert_allclose)
 from ... import units as u
 from ... import time
 from ... import coordinates as coords
+from ...units import allclose
 
 try:
     import scipy  # pylint: disable=W0611
