+++ astropy/coordinates/tests/test_frames.py
@@ -6,14 +6,15 @@ from copy import deepcopy
 import numpy as np
 
 from ... import units as u
-from ...tests.helper import (catch_warnings,
-                             pytest, quantity_allclose as allclose,
+from ...tests.helper import (catch_warnings, pytest,
                              assert_quantity_allclose as assert_allclose)
 from ...utils import OrderedDescriptorContainer
 from ...utils.compat import NUMPY_LT_1_14
 from ...utils.exceptions import AstropyWarning
 from .. import representation as r
 from ..representation import REPRESENTATION_CLASSES
+from ...units import allclose
+
 
 from .test_representation import unitphysics  # this fixture is used below
