+++ astropy/coordinates/tests/test_sky_coord.py
@@ -13,7 +13,7 @@ import numpy as np
 import numpy.testing as npt
 
 from ... import units as u
-from ...tests.helper import (catch_warnings, quantity_allclose,
+from ...tests.helper import (catch_warnings,
                              assert_quantity_allclose as assert_allclose)
 from ..representation import REPRESENTATION_CLASSES
 from ...coordinates import (ICRS, FK4, FK5, Galactic, SkyCoord, Angle,
@@ -26,6 +26,7 @@ from ...time import Time
 from ...utils import minversion, isiterable
 from ...utils.compat import NUMPY_LT_1_14
 from ...utils.exceptions import AstropyDeprecationWarning
+from ...units import allclose as quantity_allclose
 
 RA = 1.0 * u.deg
 DEC = 2.0 * u.deg
